+++ tui/timer/init.go
@@ -1,0 +1,10 @@
+package timer
+
+import tea "github.com/charmbracelet/bubbletea"
+
+// INIT
+// ----
+
+func (m timerModel) Init() tea.Cmd {
+	return nil
+}
