+++ tui/timer/model.go
@@ -1,0 +1,12 @@
+package timer
+
+// MODEL
+// -----
+
+type timerModel struct {
+	running bool // Whether the timer is running
+}
+
+func newTimerModel() timerModel {
+	return timerModel{}
+}
