+++ tui/timer/view.go
@@ -1,0 +1,11 @@
+package timer
+
+// VIEW
+// ----
+
+func (m timerModel) View() string {
+	if m.running {
+		return "Timer Running"
+	}
+	return "Timer Stopped"
+}
